Excluding tests for the page table out of  general ones for the preferences.
